feat(XCat)：可拓展自定义命令，去除重复\无用的方法
